+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Added
+
+* Adding RateLimitTrigger.request_limit (#589)
+
 ## 1.1.4 &ndash; 2026-04-20
 
 ### Added